fix: Changes for fee model (#41)
## What ❔

Changes:

1. Introduced `UPGRADE_TX_NATIVE_PER_GAS`. Upgrade transactions have 0
fee per gas and so for them the native per gas would be 0. Setting a
non-zero value there is IMO not desirable since to keep the invariant
`amount_mint >= gas_limit * fee_per_gas`. At the same time the invariant
the txs with smaller gas fee than base fee are not included is broken...
In any case, I kept the same approach as the one used on Era.
2. Previously all the withheld resources would be burned (they would get
into `native_used`
[here](https://github.com/matter-labs/zksync-os/blob/53456cbe3f3d90315b506338510cefa019ff8999/basic_bootloader/src/bootloader/process_transaction.rs#L1002-L1006).
But now I explicitly remember how much was withheld and return it. This
feature is not mandatory (firstly it seemed like mandatory, but when I
merged the `main` branch priority txs started working for me, but it is
still really nice to not waste users' funds.
3. For L1 txs I always use `transaction.max_fee_per_gas`. Today they
must be processed always, regardless of the base fee. Otherwise the
priority queue would be stalled.

## Why ❔

<!-- Why are these changes done? What goal do they contribute to? What
are the principles behind them? -->
<!-- The `Why` has to be clear to non-Matter Labs entities running their
own ZK Chain -->
<!-- Example: PR templates ensure PR reviewers, observers, and future
iterators are in context about the evolution of repos. -->

## Is this a breaking change?
- [ ] Yes
- [ ] No

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [ ] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [ ] Code has been formatted.