fix: make deployed_in_tx optional (#117)
## What ❔

<!-- What are the changes this PR brings about? -->
<!-- Example: This PR adds a PR template to the repo. -->
<!-- (For bigger PRs adding more context is appreciated) -->

Make the deployed_in_tx field of metadata an option, so that the default
case doesn't clash with accounts deployed on tx 0.

## Why ❔

<!-- Why are these changes done? What goal do they contribute to? What
are the principles behind them? -->
<!-- The `Why` has to be clear to non-Matter Labs entities running their
own ZK Chain -->
<!-- Example: PR templates ensure PR reviewers, observers, and future
iterators are in context about the evolution of repos. -->

Before, we were indexing transactions from 1, so using 0 as default to
mean "not deployed" was safe. Last week we changed it to index from 0,
meaning that now we can selfdestruct any contract on the first
transaction. This PR fixes that.

## Is this a breaking change?
- [ ] Yes
- [ ] No

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [ ] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [ ] Code has been formatted.

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>